requirements.txt
numpy==1.21.6
Flask==2.1.2
matplotlib==3.3.2
numpy==1.21.6
opencv-python-headless==4.6.0.66
pandas==1.3.5
Pillow==9.2.0
scikit_image==0.19.3
tqdm==4.60.0
seaborn==0.11.1
